# get-p-api/src/main/resources/static/oas/openapi3.yaml
openapi: 3.0.1
info:
  title: GET-P API
  description: GET-P API 명세서
  version: 0.1.0
servers:
- url: https://api.principes.xyz
  description: 개발 서버
- url: http://localhost:57020
  description: 로컬 서버
...
paths:
...
  /get-p/v2/applications/{applicationId}:
    get:
      tags:
      - 프로젝트 관리
      summary: 프로젝트 지원서 조회
      description: 의뢰자는 프로젝트 지원자의 지원서를 조회할 수 있다.
      operationId: project/get-application-form
      parameters:
      - name: applicationId
        in: path
        description: 프로젝트 지원 ID
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Bearer ${ACCESS_TOKEN}
        required: false
        schema:
          type: string
        example: Bearer ${ACCESS_TOKEN}
      responses:
        '200':
          description: '200'
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ProjectApplicationFormResponse'
              examples:
                project/get-application-form:
                  value: |-
                    {
                      "status" : 200,
                      "data" : {
                        "applicationId" : 1,
                        "type" : "TEAM",
                        "project" : {
                          "projectId" : 1,
                          "title" : "프로젝트 제목",
                          "payment" : 1000000,
                          "recruitmentCount" : 1,
                          "applicantsCount" : 5,
                          "applicationDuration" : {
                            "startDate" : "2024-07-01",
                            "endDate" : "2024-07-07"
                          },
                          "estimatedDuration" : {
                            "startDate" : "2024-07-14",
                            "endDate" : "2024-07-21"
                          },
                          "description" : "프로젝트 설명",
                          "meetingType" : "IN_PERSON",
                          "category" : "BACKEND",
                          "status" : "APPLICATION_OPENED",
                          "attachmentFiles" : [ "https://example.com/attachment1", "https://example.com/attachment2" ],
                          "hashtags" : [ "#해시태그1", "#해시태그2" ],
                          "likesCount" : 5,
                          "liked" : true,
                          "client" : {
                            "clientId" : 1,
                            "nickname" : "닉네임",
                            "address" : {
                              "zipcode" : "41566",
                              "street" : "대구광역시 북구 대학로 80",
                              "detail" : "IT대학 융복합관"
                            }
                          }
                        },
                        "nickname" : "닉네임",
                        "techStacks" : [ "Java", "Spring", "JPA" ],
                        "activityArea" : "대구광역시 달서구",
                        "education" : {
                          "school" : "경북대학교",
                          "major" : "컴퓨터공학"
                        },
                        "introduction" : "안녕하세요, 신찬규입니다.",
                        "portfolios" : [ {
                          "description" : "포트폴리오1 내용",
                          "url" : "https://github.com/scv1702/1"
                        }, {
                          "description" : "포트폴리오2 내용",
                          "url" : "https://github.com/scv1702/2"
                        } ],
                        "expectedDuration" : {
                          "startDate" : "2024-07-01",
                          "endDate" : "2024-07-31"
                        },
                        "status" : "COMPLETED",
                        "description" : "지원 내용",
                        "attachmentFiles" : [ "https://example.com/attachment1", "https://example.com/attachment2" ],
                        "teammates" : [ {
                          "peopleId" : 2,
                          "nickname" : "닉네임2",
                          "status" : "APPROVED",
                          "profileImageUrl" : "https://storage.principes.xyz/images/2/profile/image.jpg"
                        }, {
                          "peopleId" : 3,
                          "nickname" : "닉네임3",
                          "status" : "PENDING",
                          "profileImageUrl" : "https://storage.principes.xyz/images/3/profile/image.jpg"
                        }, {
                          "peopleId" : 4,
                          "nickname" : "닉네임4",
                          "status" : "APPROVED",
                          "profileImageUrl" : "https://storage.principes.xyz/images/4/profile/image.jpg"
                        } ]
                      }
                    }
  /get-p/v2/auth/login:
    post:
      tags:
      - 인증
      summary: 로그인
      description: 사용자는 로그인을 할 수 있다.
      operationId: auth/login
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              auth/login:
                value: |-
                  {
                    "email" : "test@example.com",
                    "password" : "qwer1234!"
                  }
      responses:
        '201':
          description: '201'
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              examples:
                auth/login:
                  value: |-
                    {
                      "status" : 201,
                      "data" : {
                        "grantType" : "Bearer",
                        "accessToken" : "${ACCESS_TOKEN}",
                        "refreshToken" : "${REFRESH_TOKEN}"
                      }
                    }
  /get-p/v2/auth/reissue:
    post:
      tags:
      - 인증
      summary: Access Token 재발급
      description: 사용자는 로그인 유지를 할 수 있다.
      operationId: auth/reissue-access-token
      parameters:
      - name: Refresh-Token
        in: header
        description: Bearer ${REFRESH_TOKEN}
        required: true
        schema:
          type: string
        example: Bearer ${REFRESH_TOKEN}
      responses:
        '201':
          description: '201'
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              examples:
                auth/reissue-access-token:
                  value: |-
                    {
                      "status" : 201,
                      "data" : {
                        "grantType" : "Bearer",
                        "accessToken" : "${ACCESS_TOKEN}",
                        "refreshToken" : "${REFRESH_TOKEN}"
                      }
                    }
  /get-p/v2/auth/signup:
    post:
      tags:
      - 인증
      summary: 회원 가입
      description: 사용자는 회원 가입을 할 수 있다.
      operationId: auth/signup
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
            examples:
              auth/signup:
                value: |-
                  {
                    "email" : "test@example.com",
                    "password" : "qwer1234!",
                    "verificationCode" : "1234",
                    "serviceTerms" : [ {
                      "tag" : "tag1",
                      "agreed" : true
                    }, {
                      "tag" : "tag2",
                      "agreed" : false
                    } ],
                    "memberType" : "ROLE_CLIENT"
                  }
      responses:
        '201':
          description: '201'
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                auth/signup:
                  value: |-
                    {
                      "status" : 201
                    }
...
  /get-p/v2/projects/{projectId}:
    get:
      tags:
      - 프로젝트
      summary: 프로젝트 상세 조회
      description: 사용자는 프로젝트의 상세 정보를 조회할 수 있다.
      operationId: project/get-project
      parameters:
      - name: projectId
        in: path
        description: 프로젝트 ID
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Bearer ${ACCESS_TOKEN}
        required: false
        schema:
          type: string
        example: Bearer ${ACCESS_TOKEN}
      responses:
        '200':
          description: '200'
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/DetailProjectResponse'
              examples:
                project/get-project:
                  value: |-
                    {
                      "status" : 200,
                      "data" : {
                        "projectId" : 1,
                        "title" : "프로젝트 제목",
                        "payment" : 1000000,
                        "recruitmentCount" : 1,
                        "applicantsCount" : 5,
                        "applicationDuration" : {
                          "startDate" : "2024-07-01",
                          "endDate" : "2024-07-07"
                        },
                        "estimatedDuration" : {
                          "startDate" : "2024-07-14",
                          "endDate" : "2024-07-21"
                        },
                        "description" : "프로젝트 설명",
                        "meetingType" : "IN_PERSON",
                        "category" : "BACKEND",
                        "status" : "APPLICATION_OPENED",
                        "attachmentFiles" : [ "https://example.com/attachment1", "https://example.com/attachment2" ],
                        "hashtags" : [ "#해시태그1", "#해시태그2" ],
                        "likesCount" : 5,
                        "liked" : true,
                        "client" : {
                          "clientId" : 1,
                          "nickname" : "닉네임",
                          "address" : {
                            "zipcode" : "41566",
                            "street" : "대구광역시 북구 대학로 80",
                            "detail" : "IT대학 융복합관"
                          }
                        }
                      }
                    }
...
  /get-p/v2/auth/signup/email/send:
    post:
      tags:
      - 인증
      summary: 이메일 인증 코드 전송
      description: 사용자는 회원 가입 시 이메일 인증을 해야 한다.
      operationId: auth/send-email-verification-code-for-signup
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/EmailVerificationCodeRequest'
            examples:
              auth/send-email-verification-code-for-signup:
                value: |-
                  {
                    "email" : "test@example.com"
                  }
      responses:
        '200':
          description: '200'
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                auth/send-email-verification-code-for-signup:
                  value: |-
                    {
                      "status" : 200
                    }
components:
  schemas:
    ProjectApplicationFormResponse:
      title: ProjectApplicationFormResponse
      type: object
      properties:
        data:
          type: object
          properties:
            education:
              type: object
              properties:
                major:
                  type: string
                  description: 지원자 전공
                school:
                  type: string
                  description: 지원자 학교
            portfolios:
              type: array
              items:
                type: object
                properties:
                  description:
                    type: string
                    description: 지원자 포트폴리오 설명
                  url:
                    type: string
                    description: 지원자 포트폴리오 링크
            attachmentFiles:
              type: array
              description: 첨부 파일
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            description:
              type: string
              description: 지원 내용
            project:
              type: object
              properties:
                hashtags:
                  type: array
                  description: 지원한 프로젝트 해시태그
                  items:
                    oneOf:
                    - type: object
                    - type: boolean
                    - type: string
                    - type: number
                attachmentFiles:
                  type: array
                  description: 지원한 프로젝트 첨부 파일
                  items:
                    oneOf:
                    - type: object
                    - type: boolean
                    - type: string
                    - type: number
                description:
                  type: string
                  description: 지원한 프로젝트 상세 설명
                title:
                  type: string
                  description: 지원한 프로젝트 제목
                meetingType:
                  type: string
                  description: 지원한 프로젝트 미팅 방식
                  enum:
                  - IN_PERSON
                  - REMOTE
                estimatedDuration:
                  type: object
                  properties:
                    endDate:
                      type: string
                      description: 지원한 프로젝트 예상 작업 종료 날짜
                    startDate:
                      type: string
                      description: 지원한 프로젝트 예상 작업 시작 날짜
                liked:
                  type: boolean
                  description: 지원한 프로젝트 좋아요 여부
                likesCount:
                  type: number
                  description: 지원한 프로젝트 좋아요 수
                applicationDuration:
                  type: object
                  properties:
                    endDate:
                      type: string
                      description: 지원한 프로젝트 지원자 모집 종료 날짜
                    startDate:
                      type: string
                      description: 지원한 프로젝트 지원자 모집 시작 날짜
                recruitmentCount:
                  type: number
                  description: 지원한 프로젝트 모집 인원
                applicantsCount:
                  type: number
                  description: 지원한 프로젝트 지원자 수
                client:
                  type: object
                  properties:
                    address:
                      type: object
                      properties:
                        zipcode:
                          type: string
                          description: 지원한 프로젝트 의뢰자 우편번호
                        street:
                          type: string
                          description: 지원한 프로젝트 의뢰자 도로명
                        detail:
                          type: string
                          description: 지원한 프로젝트 의뢰자 상세 주소
                    clientId:
                      type: number
                      description: 지원한 프로젝트 의뢰자 ID
                    nickname:
                      type: string
                      description: 지원한 프로젝트 의뢰자 닉네임
                payment:
                  type: number
                  description: 지원한 프로젝트 성공 보수
                category:
                  type: string
                  description: 지원한 프로젝트 카테고리
                  enum:
                  - FRONTEND
                  - BACKEND
                  - WEB
                  - PROGRAM
                  - ETC
                projectId:
                  type: number
                  description: 지원한 프로젝트 ID
                status:
                  type: string
                  description: 지원한 프로젝트 프로젝트 상태
                  enum:
                  - PREPARING
                  - APPLICATION_OPENED
                  - APPLICATION_CLOSED
                  - MEETING
                  - CONFIRMED
                  - PROGRESSING
                  - COMPLETED
                  - CANCELLED
            expectedDuration:
              type: object
              properties:
                endDate:
                  type: string
                  description: 예상 작업 종료 날짜
                startDate:
                  type: string
                  description: 예상 작업 시작 날짜
            type:
              type: string
              description: 프로젝트 지원 타입
              enum:
              - INDIVIDUAL
              - TEAM
            activityArea:
              type: string
              description: 지원자 활동 지역
            teammates:
              type: array
              description: 팀원 목록
              items:
                type: object
                properties:
                  peopleId:
                    type: number
                    description: 팀원 피플 ID
                  nickname:
                    type: string
                    description: 팀원 닉네임
                  profileImageUrl:
                    type: string
                    description: 팀원 프로필 이미지 URL
                  status:
                    type: string
                    description: 팀원 상태
                    enum:
                    - PENDING
                    - APPROVED
                    - REJECTED
                    - EXPIRED
            techStacks:
              type: array
              description: 지원자 기술 스택
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            nickname:
              type: string
              description: 지원자 닉네임
            applicationId:
              type: number
              description: 프로젝트 지원 ID
            introduction:
              type: string
              description: 지원자 자기 소개
            status:
              type: string
              description: 프로젝트 지원 상태
              enum:
              - PENDING_TEAM_APPROVAL
              - COMPLETED
              - WAITING_MEETING
              - MEETING_COMPLETED
              - ACCEPTED
              - CLOSED
        status:
          type: number
          description: 응답 상태
...
    TokenResponse:
      title: TokenResponse
      type: object
      properties:
        data:
          type: object
          properties:
            accessToken:
              type: string
              description: Access Token
            grantType:
              type: string
              description: 토큰 타입
            refreshToken:
              type: string
              description: Refresh Token
        status:
          type: number
          description: 응답 상태
...
    DetailProjectResponse:
      title: DetailProjectResponse
      type: object
      properties:
        data:
          type: object
          properties:
            hashtags:
              type: array
              description: 해시태그
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            attachmentFiles:
              type: array
              description: 첨부 파일
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            description:
              type: string
              description: 상세 설명
            title:
              type: string
              description: 제목
            meetingType:
              type: string
              description: 미팅 방식
              enum:
              - IN_PERSON
              - REMOTE
            estimatedDuration:
              type: object
              properties:
                endDate:
                  type: string
                  description: 예상 작업 종료 날짜
                startDate:
                  type: string
                  description: 예상 작업 시작 날짜
            liked:
              type: boolean
              description: 좋아요 여부
            likesCount:
              type: number
              description: 좋아요 수
            applicationDuration:
              type: object
              properties:
                endDate:
                  type: string
                  description: 지원자 모집 종료 날짜
                startDate:
                  type: string
                  description: 지원자 모집 시작 날짜
            recruitmentCount:
              type: number
              description: 모집 인원
            applicantsCount:
              type: number
              description: 지원자 수
            client:
              type: object
              properties:
                address:
                  type: object
                  properties:
                    zipcode:
                      type: string
                      description: 의뢰자 우편번호
                    street:
                      type: string
                      description: 의뢰자 도로명
                    detail:
                      type: string
                      description: 의뢰자 상세 주소
                clientId:
                  type: number
                  description: 의뢰자 ID
                nickname:
                  type: string
                  description: 의뢰자 닉네임
            payment:
              type: number
              description: 성공 보수
            category:
              type: string
              description: 카테고리
              enum:
              - FRONTEND
              - BACKEND
              - WEB
              - PROGRAM
              - ETC
            projectId:
              type: number
              description: 프로젝트 ID
            status:
              type: string
              description: 프로젝트 상태
              enum:
              - PREPARING
              - APPLICATION_OPENED
              - APPLICATION_CLOSED
              - MEETING
              - CONFIRMED
              - PROGRESSING
              - COMPLETED
              - CANCELLED
        status:
          type: number
          description: 응답 상태
    SignUpRequest:
      title: SignUpRequest
      type: object
      properties:
        password:
          type: string
          description: 비밀번호
        serviceTerms:
          type: array
          items:
            type: object
            properties:
              tag:
                type: string
                description: 서비스 약관 태그
              agreed:
                type: boolean
                description: 서비스 약관 동의 여부
        memberType:
          type: string
          description: 회원 유형. 일반 사용자는 피플인 ROLE_PEOPLE 또는 의뢰자인 ROLE_CLIENT만 가능
          enum:
          - ROLE_PEOPLE
          - ROLE_CLIENT
          - ROLE_MANAGER
          - ROLE_ADMIN
        email:
          type: string
          description: 이메일
        verificationCode:
          type: string
          description: 이메일 인증 코드
    StatusResponse:
      title: StatusResponse
      type: object
      properties:
        status:
          type: number
          description: 응답 상태
...
    LoginRequest:
      title: LoginRequest
      type: object
      properties:
        password:
          type: string
          description: 비밀번호
        email:
          type: string
          description: 이메일
...
    EmailVerificationCodeRequest:
      title: EmailVerificationCodeRequest
      type: object
      properties:
        email:
          type: string
          description: 이메일
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- JWT: [
    ]